change postMessage protocol for passing cross-origin messages
The iframe waits for the content to load, then calls postMessage to
pass a pagekey and frameid to the framed page.

Wiki listens for postMessage calls from the framed content. The
message payload is expected to take the following shape:
{
  wiki: {
    action: /* "showResult" | "doInternalLink" | "importer" | "resize" */,
    /* action-specific keys */,
    keepLineup,
    identifiers: { pagekey, frameid}
  }
}

The top-level "wiki" object is there specifically to create a
namespace for our own messages. postMessage() is a global-to-the-page
event handler. Putting our own expected messages inside a wiki
namespace should allow authors some flexibility to support messaging
other services.

This commit also introduces a resize message. This currently works
only in same-origin contexts. The intent of the resize message is to
allow framed content to expand dynamically and inform the hosting wiki
to expand accordingly.

With this addition of resize, we make the default behavior for frames
of same-origin content to resize the frame to make room to fit the
height of the framed content.

The previous protocol expected the framed page to discover the pagekey
by finding window.frameElement.name from the DOM. We learned that
window.frameElement is null in cross-origin contexts.